Add template for generic wsgi app
We have a tonn of templates in each openstack formula that describes
api wsgi app. This patch adds template to apache that allows
to create custom wsgi apps from pillar. See readme for details.
In future we swich openstack api services to use this template.

Also add `available` key to site definition. This flag is used
in conjunction with `enabled` to add config to sites-available,
but actual enablement will be done when applying specific openstack
formula.

Change-Id: I008e00958f4ddb36735b6399cbc5d2957b893889